+++ resources/views/reference_users/index.blade.php
@@ -9,11 +9,13 @@
         <div class="d-sm-flex align-items-center justify-content-between mb-4">
             <h1 class="h3 mb-0 text-gray-800">All Reference User</h1>
             <div class="row">
-                <div class="col-md-6">
-                    <a href="{{ route('reference_users.create') }}" class="btn btn-sm btn-primary">
-                        <i class="fas fa-plus"></i> Add New
-                    </a>
-                </div>
+                @if (auth()->user()->hasPermissionTo('reference-user-create'))
+                    <div class="col-md-6">
+                        <a href="{{ route('reference_users.create') }}" class="btn btn-sm btn-primary">
+                            <i class="fas fa-plus"></i> Add New
+                        </a>
+                    </div>
+                @endif
                 <div class="col-md-6">
                     <a href="{{ route('reference_users.export') }}" class="btn btn-sm btn-success">
                         <i class="fas fa-check"></i> Export To Excel
@@ -70,25 +72,31 @@
                                         @endif
                                     </td>
                                     <td style="display: flex">
-                                        @if ($reference_user->status == 0)
-                                            <a href="{{ route('reference_users.status', ['reference_user_id' => $reference_user->id, 'status' => 1]) }}"
-                                                class="btn btn-success m-2">
-                                                <i class="fa fa-check"></i>
-                                            </a>
-                                        @elseif ($reference_user->status == 1)
-                                            <a href="{{ route('reference_users.status', ['reference_user_id' => $reference_user->id, 'status' => 0]) }}"
-                                                class="btn btn-danger m-2">
-                                                <i class="fa fa-ban"></i>
+                                        @if (auth()->user()->hasPermissionTo('reference-user-delete'))
+                                            @if ($reference_user->status == 0)
+                                                <a href="{{ route('reference_users.status', ['reference_user_id' => $reference_user->id, 'status' => 1]) }}"
+                                                    class="btn btn-success m-2">
+                                                    <i class="fa fa-check"></i>
+                                                </a>
+                                            @elseif ($reference_user->status == 1)
+                                                <a href="{{ route('reference_users.status', ['reference_user_id' => $reference_user->id, 'status' => 0]) }}"
+                                                    class="btn btn-danger m-2">
+                                                    <i class="fa fa-ban"></i>
+                                                </a>
+                                            @endif
+                                        @endif
+                                        @if (auth()->user()->hasPermissionTo('reference-user-edit'))
+                                            <a href="{{ route('reference_users.edit', ['reference_user' => $reference_user->id]) }}"
+                                                class="btn btn-primary m-2">
+                                                <i class="fa fa-pen"></i>
                                             </a>
                                         @endif
-                                        <a href="{{ route('reference_users.edit', ['reference_user' => $reference_user->id]) }}"
-                                            class="btn btn-primary m-2">
-                                            <i class="fa fa-pen"></i>
-                                        </a>
 
-                                        <a class="btn btn-danger m-2" href="javascript:void(0);"
-                                            onclick="delete_conf_common('{{ $reference_user->id }}','ReferenceUser', 'Reference User', '{{ route('reference_users.index') }}');"><i
-                                                class="fas fa-trash"></i></a>
+                                        @if (auth()->user()->hasPermissionTo('reference-user-delete'))
+                                            <a class="btn btn-danger m-2" href="javascript:void(0);"
+                                                onclick="delete_conf_common('{{ $reference_user->id }}','ReferenceUser', 'Reference User', '{{ route('reference_users.index') }}');"><i
+                                                    class="fas fa-trash"></i></a>
+                                        @endif
                                     </td>
                                 </tr>
                             @empty
